deep cuts as -o parameter. gave up on multiple commands because i couldn't get default choice to work.
